+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add `--force` CLI flag, and do not overwrite files by default.
 - Improve output file pathing.
 - Fix password file setting in Puppet bindings.
+- Fix longform output directory CLI argument.
 
 ### 1.3.1
 - Fix GPG keys output path message.